updating the inpainting code v3

+++ inpainting_losses.py
@@ -281,9 +281,15 @@ class InpaintingLoss(nn.Module):
             target_dilated = (target_dilated > 0).float()  # (B,1,H,W)
 
         # BCE between dilated maps
+        # F.binary_cross_entropy is blocked under autocast (float16 BCE is
+        # numerically unsafe — PyTorch raises a RuntimeError by design).
+        # Upcasting to float32 for this one op is the standard fix; gradients
+        # flow correctly back into the float16 graph via autocast bookkeeping.
         pred_dilated_clamped = pred_dilated.clamp(1e-6, 1.0 - 1e-6)
-        loss = F.binary_cross_entropy(pred_dilated_clamped,
-                                      target_dilated.detach())
+        loss = F.binary_cross_entropy(
+            pred_dilated_clamped.float(),
+            target_dilated.detach().float()
+        )
         return loss
 
     # ── Combined Forward ──────────────────────────────────────────────────────
